extract loading spinner component to a shared location
